log buttons

+++ frontend/src/components/Merchant/Navigation/index.js
@@ -17,23 +17,36 @@ class Navigation extends Component {
     }
 
     handleLogOutButton() {
-        const {history} = this.props;
-        let headers = {
-            'authorization': JSON.parse(localStorage.activeUser)['token'] 
-        };        
-        let config = {
-            'headers': headers
-        };
-        
-        axios.get(`${settings.API_ROOT}/logout`, config)
-            .then(response => {
-                console.log(response);
-                localStorage.setItem('activeUser', '');
-                history.push('/login');
-            })
-            .catch(error => {
-                console.log(error);                
-            })
+        swal({
+            title: "Log out?",
+            text: "Are you sure you want log out?",            
+            type: "warning",
+            buttons: {
+                confirm:true,
+                cancel: true
+            }            
+          })
+          .then((confirm) => {   
+              if(confirm){
+                const {history} = this.props;
+                let headers = {
+                    'authorization': JSON.parse(localStorage.activeUser)['token'] 
+                };        
+                let config = {
+                    'headers': headers
+                };
+                
+                axios.get(`${settings.API_ROOT}/logout`, config)
+                    .then(response => {
+                        console.log(response);
+                        localStorage.setItem('activeUser', '');
+                        this.forceUpdate();
+                    })
+                    .catch(error => {
+                        console.log(error);                
+                    })
+                };
+            });
     }
 
     handleDeleteAccount() {
@@ -70,23 +83,28 @@ class Navigation extends Component {
                     // })
                 }                
             });
-            
-            
+    }
 
+    renderLogBtn(){
+        if(localStorage.activeUser){
+            return (
+                <Dropdown text='Account' className='ui primary button'>
+                    <Dropdown.Menu>
+                        <Dropdown.Item onClick={() => this.handleLogOutButton()}>Logout</Dropdown.Item>
+                        <Dropdown.Divider />
+                        <Dropdown.Item onClick={() => this.handleDeleteAccount()}>Delete Account</Dropdown.Item>
+                    </Dropdown.Menu>
+                </Dropdown>
+            )
+        } else {
+            return (
+                <Link to={`/login`} className="item">Login</Link>
+            )
+        }
         
     }
 
     render() {
-        const options = ( 
-              <Dropdown text='Options' className='ui primary button'>
-                <Dropdown.Menu>
-                  <Dropdown.Item onClick={() => this.handleLogOutButton()}>Logout</Dropdown.Item>
-                  <Dropdown.Divider />
-                  <Dropdown.Item onClick={() => this.handleDeleteAccount()}>Delete Account</Dropdown.Item>
-                </Dropdown.Menu>
-              </Dropdown>
-        );
-          
 
         return (    
             <div>
@@ -97,7 +115,6 @@ class Navigation extends Component {
                 </Link>
 
                 <Link to={`/`} className="item">Catalog</Link>
-
 
             <div className="right menu">
 
@@ -111,7 +128,7 @@ class Navigation extends Component {
                     <i className="shopping basket icon mx-auto"></i>
                 </Link>
                 <div className="item">
-                    {options}
+                    {this.renderLogBtn()}
                 </div>
             </div>
         </div>
